// src/communication/models/get-game-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * QuizApi
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 *
 * @export
 * @interface GetGameResponse
 */
export interface GetGameResponse {
  /**
   *
   * @type {string}
   * @memberof GetGameResponse
   */
  id?: any;
  /**
   *
   * @type {BaseUserModel}
   * @memberof GetGameResponse
   */
  user1?: any;
  /**
   *
   * @type {BaseUserModel}
   * @memberof GetGameResponse
   */
  user2?: any;
  /**
   *
   * @type {boolean}
   * @memberof GetGameResponse
   */
  isDone?: any;
  /**
   *
   * @type {Array<QuestionModel>}
   * @memberof GetGameResponse
   */
  user1Result?: any[] | null;
  /**
   *
   * @type {Array<QuestionModel>}
   * @memberof GetGameResponse
   */
  user2Result?: any[] | null;
}